Update 1.0.7.0
- Switched to " (double quotation marks) instead of ' (single quotation marks) to stay consistent with Roblox code

- Switched out coroutine.resume for task.spawn, which does not break continuations.

- GetPing is no longer divided by two, as ping is a two-way trip rather than one-way.

General tip: if you're calculating your hitbox on the client, please also validate on whether the hit-position is accurate, as exploiters could spoof this value whenever it's being sent over to the server. There is no one reliable way to verify this, and as such you should program your own solutions depending on the context.